+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 - Integrate with memory ([#809](https://github.com/opensearch-project/dashboards-flow-framework/pull/809))
 - Add version filtering on agentic search usecase ([#813](https://github.com/opensearch-project/dashboards-flow-framework/pull/813))
 - Improve 'Visualized Hits' values ([#814](https://github.com/opensearch-project/dashboards-flow-framework/pull/814))
+- Automatically add response filters to flow agents when possible ([#817](https://github.com/opensearch-project/dashboards-flow-framework/pull/817))
 ### Bug Fixes
 ### Infrastructure
 ### Documentation